Recognizes python and java files trash aswell now, along with cpp trash files
Because I selected the "C++" template when I created this repo, GitHub automatically generated a .gitignore file that only knows how to look for C++ junk. It won't realize that my Java .class files or Python __pycache__ folders are trash. So I updated this file such that it'll recognize those as trash aswell.